--- src/navigationBanner/archivepage/ArchivePage.css
@@ -1,35 +1,0 @@
-#title {
-    text-align: center;
-    font-family: arial, sans-serif;
-  }
- 
-  
-  
-  #Archived_Items {
-    margin-left: auto;
-    margin-right:auto;
-    text-align: center;
-    font-family: "Trebuchet MS", Arial, Helvetica, sans-serif;
-    border-collapse: collapse;
-    border: 3px solid #ddd;
-    width: 70%;
-  }
-  
-  #Archived_Items td, #Archived_Items th {
-    border: 1px solid #ddd;
-    padding: 8px;
-    
-  }
-  
-  #Archived_Items tr:nth-child(even){background-color: #f2f2f2;}
-  
-  #Archived_Items tr:hover {background-color: #ddd;}
-  
-  #Archived_Items th {
-    padding-top: 12px;
-    padding-bottom: 12px;
-    text-align: center;
-    background-color: #0e7e92;
-    color: white;
-  }
-
